+++ tests/testthat/test-dataframe.R
@@ -1327,3 +1327,34 @@ test_that("rolling for DataFrame: can be ungrouped", {
     to_data_frame()
   expect_equal(actual, df$to_data_frame())
 })
+
+test_that("flags work", {
+  df = pl$DataFrame(a = c(2, 1), b = c(3, 4), c = list(c(1, 2), 4))
+  expect_identical(
+    df$sort("a")$flags,
+    list(
+      a = list(SORTED_ASC = TRUE, SORTED_DESC = FALSE),
+      b = list(SORTED_ASC = FALSE, SORTED_DESC = FALSE),
+      c = list(SORTED_ASC = FALSE, SORTED_DESC = FALSE, FAST_EXPLODE = FALSE)
+    )
+  )
+  expect_identical(
+    df$with_columns(pl$col("b")$implode())$flags,
+    list(
+      a = list(SORTED_ASC = FALSE, SORTED_DESC = FALSE),
+      b = list(SORTED_ASC = FALSE, SORTED_DESC = FALSE, FAST_EXPLODE = TRUE),
+      c = list(SORTED_ASC = FALSE, SORTED_DESC = FALSE, FAST_EXPLODE = TRUE)
+    )
+  )
+
+  df = pl$DataFrame(
+    a = list(c(1, 2), c(4, 5)),
+    schema = list(a = pl$Array(pl$Float64, 2))
+  )
+  expect_identical(
+    df$flags,
+    list(
+      a = list(SORTED_ASC = FALSE, SORTED_DESC = FALSE, FAST_EXPLODE = FALSE)
+    )
+  )
+})